+++ netket/experimental/observable/variance/exact.py
@@ -18,7 +18,6 @@ from functools import partial
 
 from netket.vqs import FullSumState, expect, expect_and_grad
 import netket.jax as nkjax
-from netket.utils import mpi
 from netket.stats import Stats
 
 from .variance_operator import VarianceObservable
@@ -89,6 +88,5 @@ def expect_and_grad_inner_fs(
     var, var_vjp_fun = nkjax.vjp(expect_kernel_var, params, conjugate=True)
 
     var_grad = var_vjp_fun(jnp.ones_like(var))[0]
-    var_grad = jax.tree_util.tree_map(lambda x: mpi.mpi_mean_jax(x)[0], var_grad)
 
     return Stats(mean=var, error_of_mean=0.0, variance=0.0), var_grad