+++ netket/utils/struct/pytree_serialization_sharding.py
@@ -30,17 +30,6 @@ class ShardedFieldSpec:
     fields are ignored.
     """
 
-    mpi_sharded_axis: int | None = 0
-    """
-    Optional integer indicating which axis is to be considered sharded when running with
-    MPI. Defaults to 0.
-
-    If this is specified, loading with MPI a state saved from a run with jax sharding will
-    scatter/partition the data along the specified axis.
-
-    If None, MPI is not supported.
-    """
-
     deserialization_function: ShardedDeserializationFunction | str | None = "relaxed"
     """
     Function to use to deserialize the data. Can be a callable with the signature:
@@ -107,16 +96,11 @@ class ShardedFieldSpec:
 
 
 def to_flax_state_dict_sharding(sharded_data):
-    from netket.utils import mpi
-
     # This exists for backward compatibility...
     from netket.jax.sharding import gather
     from netket import config as nkconfig
 
-    if mpi.n_nodes > 1:
-        gathered_data = mpi.mpi_gather(sharded_data)
-        result = gathered_data.reshape((-1,) + gathered_data.shape[2:])
-    elif (
+    if (
         nkconfig.netket_experimental_sharding
         and not nkconfig.netket_experimental_sharding_fast_serialization
     ):
@@ -139,37 +123,10 @@ def from_flax_state_dict_sharding_fail(value_target, value_state, *, name="."):
 
 # Strict mode
 def from_flax_state_dict_sharding_strict(value_target, value_state, *, name="."):
-    # sharding imports
-    from netket.utils import mpi
-
-    sharded_dim = 0
-
-    if mpi.n_nodes > 1:
-        # Assuming every rank got the full data... So we don't need to broadcast
-        if (
-            value_state.shape[sharded_dim]
-            != mpi.n_nodes * value_target.shape[sharded_dim]
-        ):
-            raise ValueError(
-                f"Sharded data ({name}) has shape {value_state.shape} but target has shape {value_target.shape}."
-                "This probably means that you serialized the data with a different number of nodes or a different "
-                "array size alltogether."
-            )
-        scattered_value_state = value_state.reshape(
-            (mpi.n_nodes,) + value_target.shape
-        )[mpi.rank]
-        result = serialization.from_state_dict(
-            value_target, scattered_value_state, name=name
-        )
-        result = jax.lax.with_sharding_constraint(result, value_target.sharding)
-
-    else:  # if processes > 1
-        unsharded_update = serialization.from_state_dict(
-            value_target, value_state, name=name
-        )
-        result = jax.lax.with_sharding_constraint(
-            unsharded_update, value_target.sharding
-        )
+    unsharded_update = serialization.from_state_dict(
+        value_target, value_state, name=name
+    )
+    result = jax.lax.with_sharding_constraint(unsharded_update, value_target.sharding)
     return result
 
 
@@ -177,66 +134,46 @@ def from_flax_state_dict_sharding_relaxed(
     value_target, value_state, *, name=".", ignore_errors: bool = False
 ):
     # sharding imports
-    from netket.utils import mpi
-
     sharded_dim = 0
 
     serialized_total_sharded_size = (
         value_state.shape[sharded_dim] if value_state.ndim > 0 else 0
     )
     target_total_sharded_size = (
-        mpi.n_nodes * value_target.shape[sharded_dim]
-        if value_target.ndim > 0
-        else mpi.n_nodes
+        value_target.shape[sharded_dim] if value_target.ndim > 0 else 1
     )
 
     if serialized_total_sharded_size > target_total_sharded_size:
-        if mpi.rank == 0 and jax.process_index() == 0:
+        if jax.process_index() == 0:
             warnings.warn(
                 f"Sharded data ({name}) has shape {value_state.shape} but target has shape {value_target.shape}."
                 "This probably means that you serialized the data with more nodes. We will try to load it anyway "
                 "by ignoring the extra data."
             )
         if value_target.ndim == 0:
-            value_state = value_state[: mpi.n_nodes]
+            value_state = value_state[:1]
         else:
-            value_state = value_state[: mpi.n_nodes * value_target.shape[sharded_dim]]
+            value_state = value_state[: value_target.shape[sharded_dim]]
     elif serialized_total_sharded_size < target_total_sharded_size:
         if ignore_errors:
             return value_target
         raise ValueError(
             f"\n"
-            f"Sharded (or mpi distributed) data for field '{name}' size mismatch: \n"
+            f"Sharded data for field '{name}' size mismatch: \n"
             f"\t Info                                          \n"
             f"\t       sharded axis #           : {sharded_dim}\n"
-            f"\t       # mpi nodes              : {mpi.n_nodes}\n"
             f"\t       # jax processes          : {jax.process_count()}\n"
             f"\t serialized shape   (global)    : {value_state.shape}\n"
             f"\t           (total_sharded_size) : {serialized_total_sharded_size}\n"
-            f"\t restore trgt shape (global)    : {(mpi.n_nodes,) + value_target.shape}\n"
-            f"\t                    (per-shard) : {value_target.shape}\n"
+            f"\t restore trgt shape (global)    : {value_target.shape}\n"
             f"\t           (total_sharded_size) : {target_total_sharded_size}\n"
             "This probably means that you serialized the data with less nodes. We cannot load the data in this case."
             "\n"
         )
-
-    if mpi.n_nodes > 1:
-        # Assuming every rank got the full data... So we don't need to broadcast
-        scattered_value_state = value_state.reshape(
-            (mpi.n_nodes,) + value_target.shape
-        )[mpi.rank]
-        result = serialization.from_state_dict(
-            value_target, scattered_value_state, name=name
-        )
-        result = jax.lax.with_sharding_constraint(result, value_target.sharding)
-
-    else:  # if processes > 1
-        unsharded_update = serialization.from_state_dict(
-            value_target, value_state, name=name
-        )
-        result = jax.lax.with_sharding_constraint(
-            unsharded_update, value_target.sharding
-        )
+    unsharded_update = serialization.from_state_dict(
+        value_target, value_state, name=name
+    )
+    result = jax.lax.with_sharding_constraint(unsharded_update, value_target.sharding)
     return result
 
 
@@ -249,21 +186,17 @@ def from_flax_state_dict_sharding_relaxed_rng_key(
     is larger than the serialized state.
     """
     # sharding imports
-    from netket.utils import mpi
-
     sharded_dim = 0
 
     serialized_total_sharded_size = (
         value_state.shape[sharded_dim] if value_state.ndim > 0 else 0
     )
     target_total_sharded_size = (
-        mpi.n_nodes * value_target.shape[sharded_dim]
-        if value_target.ndim > 0
-        else mpi.n_nodes
+        value_target.shape[sharded_dim] if value_target.ndim > 0 else 1
     )
 
     if serialized_total_sharded_size > target_total_sharded_size:
-        if mpi.rank == 0 and jax.process_index() == 0:
+        if jax.process_index() == 0:
             warnings.warn(
                 f"""
                 Serialized RNG key data ({name}) has a BIGGER shape {value_state.shape} that the
@@ -276,11 +209,11 @@ def from_flax_state_dict_sharding_relaxed_rng_key(
                 """
             )
         if value_target.ndim == 0:
-            value_state = value_state[: mpi.n_nodes]
+            value_state = value_state[:1]
         else:
-            value_state = value_state[: mpi.n_nodes * value_target.shape[sharded_dim]]
+            value_state = value_state[: value_target.shape[sharded_dim]]
     elif serialized_total_sharded_size < target_total_sharded_size:
-        if mpi.rank == 0 and jax.process_index() == 0:
+        if jax.process_index() == 0:
             warnings.warn(
                 f"""
                 Serialized RNG key data ({name}) has a SMALLER shape {value_state.shape} that the
@@ -293,10 +226,8 @@ def from_flax_state_dict_sharding_relaxed_rng_key(
                 If it fails, open an issue.
                 """
             )
-        # construct the taget shape that we really-really want by multiplying the target
-        # sharded dimension by number of mpi ranks if available
         target_shape = list(value_target.shape)
-        target_shape[sharded_dim] = mpi.n_nodes * value_target.shape[sharded_dim]
+        target_shape[sharded_dim] = value_target.shape[sharded_dim]
         target_shape = tuple(target_shape)
 
         # Generate a new RNG key data with the target shape, using the current rng data as
@@ -312,21 +243,8 @@ def from_flax_state_dict_sharding_relaxed_rng_key(
         new_value_state[: value_state.shape[0]] = value_state
         value_state = new_value_state
 
-    if mpi.n_nodes > 1:
-        # Assuming every rank got the full data... So we don't need to broadcast
-        scattered_value_state = value_state.reshape(
-            (mpi.n_nodes,) + value_target.shape
-        )[mpi.rank]
-        result = serialization.from_state_dict(
-            value_target, scattered_value_state, name=name
-        )
-        result = jax.lax.with_sharding_constraint(result, value_target.sharding)
-
-    else:  # if processes > 1
-        unsharded_update = serialization.from_state_dict(
-            value_target, value_state, name=name
-        )
-        result = jax.lax.with_sharding_constraint(
-            unsharded_update, value_target.sharding
-        )
+    unsharded_update = serialization.from_state_dict(
+        value_target, value_state, name=name
+    )
+    result = jax.lax.with_sharding_constraint(unsharded_update, value_target.sharding)
     return result