+++ test/optimizer/test_qgt_logic.py
@@ -36,7 +36,7 @@ from netket.optimizer.qgt import (
     qgt_jacobian_common,
 )
 import netket as nk
-from netket.jax.sharding import distribute_to_devices_along_axis, device_count_per_rank
+from netket.jax.sharding import distribute_to_devices_along_axis
 
 from .. import common
 
@@ -265,7 +265,7 @@ def test_reassemble_complex(e):
 
 
 @common.named_parametrize("holomorphic", [True, False])
-@common.named_parametrize("n_samp", [24 * device_count_per_rank(), 1024])
+@common.named_parametrize("n_samp", [24 * jax.device_count(), 1024])
 @common.named_parametrize("jit", [True, False])
 @pytest.mark.parametrize("outdtype, pardtype", all_test_types)
 @common.named_parametrize("chunk_size", [8, None])
@@ -296,7 +296,7 @@ def test_matvec(e, jit, chunk_size):
 
 
 @common.named_parametrize("holomorphic", [True, False])
-@common.named_parametrize("n_samp", [24 * device_count_per_rank(), 1024])
+@common.named_parametrize("n_samp", [24 * jax.device_count(), 1024])
 @common.named_parametrize("jit", [True, False])
 @pytest.mark.parametrize("outdtype, pardtype", all_test_types)
 @common.named_parametrize("chunk_size", [8, None])
@@ -342,7 +342,7 @@ def test_matvec_linear_transpose(e, jit, chunk_size):
 
 # TODO separate test for prepare_centered_oks
 @common.named_parametrize("holomorphic", [True])
-@common.named_parametrize("n_samp", [25 * device_count_per_rank(), 1024])
+@common.named_parametrize("n_samp", [25 * jax.device_count(), 1024])
 @common.named_parametrize("jit", [True, False])
 @common.named_parametrize("chunk_size", [7, None])
 @pytest.mark.parametrize(
@@ -379,7 +379,7 @@ def test_matvec_treemv(e, jit, holomorphic, pardtype, outdtype, chunk_size):
 # TODO separate test for prepare_centered_oks
 # TODO test C->R ?
 @common.named_parametrize("holomorphic", [True, False])
-@common.named_parametrize("n_samp", [25 * device_count_per_rank(), 1024])
+@common.named_parametrize("n_samp", [25 * jax.device_count(), 1024])
 @common.named_parametrize("jit", [True, False])
 @pytest.mark.parametrize("outdtype, pardtype", test_types)
 def test_matvec_treemv_modes(e, jit, holomorphic, pardtype, outdtype):
@@ -435,7 +435,7 @@ def e_offset(n_samp, outdtype, pardtype, holomorphic, offset, seed=123):
 
 
 @pytest.mark.parametrize("holomorphic", [True])
-@pytest.mark.parametrize("n_samp", [25 * device_count_per_rank(), 1024])
+@pytest.mark.parametrize("n_samp", [25 * jax.device_count(), 1024])
 @pytest.mark.parametrize(
     "outdtype, pardtype",
     r_c_test_types,  # r_r_test_types + c_c_test_types + r_c_test_types